requirements.txt
appnope==0.1.2
asgiref==3.3.1
attrs==20.3.0
autopep8==1.5.5
backcall==0.2.0
certifi==2020.12.5
cffi==1.14.5
chardet==4.0.0
cryptography==3.4.7
decorator==4.4.2
defusedxml==0.7.1
Django==3.1.7
django-allauth==0.44.0
django-model-utils==4.1.1
django-notifications-hq==1.6.0
django-sslserver==0.22
idna==2.10
iniconfig==1.1.1
jedi==0.18.0
jsonfield==3.1.0
oauthlib==3.1.0
packaging==20.9
parso==0.8.2
pexpect==4.8.0
pickleshare==0.7.5
pluggy==0.13.1
prompt-toolkit==3.0.18
ptyprocess==0.7.0
py==1.10.0
pycodestyle==2.6.0
pycparser==2.20
Pygments==2.8.1
PyJWT==2.0.1
pyparsing==2.4.7
pytest==6.2.2
python-dateutil==2.8.1
python3-openid==3.2.0
pytz==2021.1
pyzmq==22.0.3
requests==2.25.1
requests-oauthlib==1.3.0
six==1.15.0
sqlparse==0.4.1
swapper==1.1.2.post1
toml==0.10.2
tornado==6.1
traitlets==5.0.5
typing-extensions==3.7.4.3
urllib3==1.26.4
wcwidth==0.2.5
